Fix/division algs fixes (#158)
* Fixed updateQueuePos and no maxParticipants bugs, added new rows to course factory

* Rework of GroupCourseDivision + some bugfixes

* Added missing comment to GroupCourseDivision

* Fixed rework of GCD

* Refactored Divisions to remove duplicate code and fixed rework of GBD

* fix linting

* Fix/division (#159)

* fix array position for non drinkers

* remove redeclared abstract functions

* get max_groups and max_participants for balanced division

* set queue pos if event has no slots

* parse max_groups and max_participants to division alg
